split model into separate files
